src/admin-panel/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.1.0
django-filter==23.5
markdown==3.5.2
drf-yasg==1.21.7
whitenoise==6.6.0
python-dotenv==1.0.1
gdal==3.6.4
django-cors-headers==4.3.1
django-allauth==0.61.0 